alita_agent_prototype/requirements.txt
# Core async and web libraries
asyncio>=3.4.3
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0


# Optional API clients (install as needed)
openai>=1.0.0
google-generativeai>=0.5.0
# anthropic>=0.20.0

# Development tools
pytest>=7.2.0
pytest-asyncio>=0.20.0
black>=23.0.0
flake8>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
